Move startup and argument parsing to trollflow2.launcher module
